cacheword now initialized with default pin in first run activity so that sign up process can connect to the db
